upgrade float32 to float64

+++ ftx/get_candles.go
@@ -11,11 +11,11 @@ import (
 type CandleData struct {
 	StartTime string	`json:"startTime"`
 	Time float64	`json:"time"`
-	Open float32	`json:"open"`
-	High float32	`json:"high"`
-	Low float32		`json:"low"`
-	Close float32	`json:"close"`
-	Volume float32	`json:"volume"`
+	Open float64	`json:"open"`
+	High float64	`json:"high"`
+	Low float64		`json:"low"`
+	Close float64	`json:"close"`
+	Volume float64	`json:"volume"`
 }
 
 type CandleDataResult struct {
